fix(gha): workaround cache-to and cache-from scope
This commit introduces a workaround for the GitHub Actions cache backend
when using the docker `cache-from` and `cache-to` flags with BuildKit.

The current version of `int128/docker-build-cache-config-action` incorrectly uses
`ref=` instead of `scope=` for the `gha` cache backend, which leads to shared
caches across branches and builds, resulting in frequent invalidations and
inefficient caching.

Until the upstream PR is merged and released:
https://github.com/int128/docker-build-cache-config-action/pull/1213

...this commit replace `ref=` by `scope=` to ensure proper cache
isolation and reliable cache behavior in CI.

This workaround will be removed once the upstream fix is available in a stable release.